Added pellets that can be collected to increase score
Added pellets that can be collected to increase score.
Moved PlayerBeetle widget to a character.py.
Moved Level widget to level.py.
