+++ Hotrod/collectable.py
@@ -1,0 +1,10 @@
+__author__ = 'Harriet'
+
+from kivy.uix.widget import Widget
+from kivy.properties import NumericProperty
+from kivy.properties import ReferenceListProperty
+
+class Pellet(Widget):
+    coordinates_x = NumericProperty(0)
+    coordinates_y = NumericProperty(0)
+    coordinates = ReferenceListProperty(coordinates_x, coordinates_y)
